Turn goose entrypoint into a library function (#1753)
