+++ crates/goose-cli/src/lib.rs
@@ -1,5 +1,6 @@
 use etcetera::AppStrategyArgs;
 use once_cell::sync::Lazy;
+pub mod cli;
 pub mod commands;
 pub mod logging;
 pub mod session;
